Add local conftest.py
